+++ resources/js/app.js
@@ -35,10 +35,13 @@ Vue.component('user-collections-nav-link', require('./components/user-collection
 Vue.component('user-collections-favourite-button', require('./components/user-collections/FavouriteButton.vue').default);
 Vue.component('user-collections-clear-button', require('./components/user-collections/ClearButton.vue').default);
 Vue.component('user-collections-share-form', require('./components/user-collections/ShareForm.vue').default);
+Vue.component('user-collections-store', require('./components/user-collections/Store.vue').default);
+Vue.component('user-collections-shared-collection', require('./components/user-collections/SharedCollection.vue').default);
 Vue.component('color-widget', require('./components/ColorWidget.vue').default);
 Vue.component('year-slider', require('./components/YearSlider.vue').default);
 Vue.component('inline-input', require('./components/InlineInput.vue').default);
-Vue.component('copy-to-clipboard', require('./components/CopyToClipboard.vue').default);
+Vue.component('copy-to-clipboard-group', require('./components/CopyToClipboardGroup.vue').default);
+Vue.component('copy-to-clipboard-link', require('./components/CopyToClipboardLink.vue').default);
 
 Vue.component('slider', require('./components/vue/slider').default);
 Vue.component('color-slider', require('./components/vue/color-slider').default);