+++ resources/lang/en/user-collection.php
@@ -3,11 +3,10 @@
 return array(
     'title'            => 'Favourite artworks',
     'empty'            => 'You have no saved favourite artworks.',
-    'content-intro'    => 'Here you will find the artworks you have saved as favourites. <br>
-The artworks are only temporarily stored in your browser.',
-    'content-usage'    => 'Click on the star to remove the work from the list. <br>
-Deleted works will disappear when the page is reloaded.',
-    'content-feedback' => 'You can send bugs, comments and suggestions to <a href="mailto:lab@sng.sk"> lab@sng.sk </a>',
+    'content-intro'    => 'Your favourite artworks are only stored in this browser.',
+    'content-usage'    => 'Click on the star to remove the work from the list. Removed artworks will disappear once you reload the page.',
+    'share-info'       => 'You can create share your favourites as a selection.',
     'clear'            => 'delete all artworks',
     'clear-confirm'    => 'Are you sure you want to delete all your favourite artworks from the list?',
+    'share'            => 'create new selection',
 );